requirements.txt
fastapi
uvicorn[standard]
python-jose[cryptography]
requests
pytest
httpx
sentence-transformers
faiss-cpu
numpy
